Take string template from caller
